#47 feat: efficiency on deleting

+++ app/api/appointment/timeslot/delete/[timeslotId]/route.ts
@@ -1,4 +1,4 @@
-import { Prisma } from "@prisma/client";
+import { Prisma, TimeSlotStatus } from "@prisma/client";
 import { NextResponse } from "next/server";
 import { prisma } from "@/prisma";
 import { MISSING_PARAMETERS } from "@/utils/constants";
@@ -13,6 +13,16 @@ export async function DELETE(request: Request) {
             return NextResponse.json({ message: MISSING_PARAMETERS }, { status: 400 });
         }
 
+        const existingTimeslot = await prisma.timeSlot.findFirst({
+            where: { id: timeSlotId }, include: { appointment: true }
+        })
+        if (!existingTimeslot) {
+            return NextResponse.json({ message: "Timeslot doesn't exist!" }, { status: 400 });
+        }
+
+        if (existingTimeslot?.appointment) {
+            return NextResponse.json({ message: "Can't delete because there is an associated appoiontment" }, { status: 400 });
+        }
         await prisma.timeSlot.delete({ where: { id: timeSlotId } })
 
         return NextResponse.json(
